[llm_d] toolbox: ensure_gateway: moved from the kserve project
