+++ projects/llm_d/toolbox/ensure_gateway/templates/gateway.yaml.j2
